int not selected in corr matrix

+++ 01-data-preprocessing/correlation_heatmap.py
@@ -7,7 +7,7 @@ df = df.dropna(subset=["VOTES", "Gross"])
 df["VOTES"] = df["VOTES"].str.replace(",", "").astype(int)
 df["Gross"] = (df["Gross"].str.replace("$", "")).str.replace("M", "").astype(float)
 
-numeric_df = df.select_dtypes(include=["float64", "int64"])
+numeric_df = df.select_dtypes(include=["float", "int"])
 corr_matrix = numeric_df.corr()
 
 sns.heatmap(corr_matrix, annot=True, fmt=".2f")
